refactor(song-management): #54  move string trait
