+++ src/Shared/Domain/FirstName.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\Shared\Domain;
 
-use App\Shared\StringValueObjectInterface;
-
 final readonly class FirstName
 {
     use StringValueObjectInterface;
