+++ src/Shared/Domain/StringValueObjectInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Shared;
+namespace App\Shared\Domain;
 
 trait StringValueObjectInterface
 {
