+++ src/Shared/Domain/User/User.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Shared\Domain\User;
+
+readonly class User
+{
+    private function __construct(
+    ) {
+    }
+}
