+++ src/SongManagement/Domain/Model/Contributor/Biography.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\SongManagement\Domain\Model\Contributor;
 
-use App\Shared\StringValueObjectInterface;
+use App\Shared\Domain\StringValueObjectInterface;
 
 readonly class Biography
 {
