+++ src/SongManagement/Domain/Model/Song/Title.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\SongManagement\Domain\Model\Song;
 
-use App\Shared\StringValueObjectInterface;
+use App\Shared\Domain\StringValueObjectInterface;
 
 /**
  * Title Value Object.
